pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "beeg_api"
version = "1.0"
description = "A Python API for the Porn Site beeg.com"
readme = "README.md"
requires-python = ">=3.6"
license = { text = "LGPLv3" }
authors = [
    { name = "Johannes Habel", email = "EchterAlsFake@proton.me" }
]
dependencies = [
    "eaf_base_api",
    "httpx",
    "h2"
]
optional-dependencies = {av = ["av"], ffmpeg = ["ffmpeg-progress-yield"]}
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python",
]

[project.urls]
Homepage = "https://github.com/EchterAlsFake/beeg_api"


[tool.setuptools]
packages = { find = {} }